Add initial implementation of KeyConfiguration overlay
