+++ osu.Game/Overlays/KeyConfiguration/KeyConfiguration.cs
@@ -1,0 +1,30 @@
+// Copyright (c) 2007-2017 ppy Pty Ltd <contact@ppy.sh>.
+// Licensed under the MIT Licence - https://raw.githubusercontent.com/ppy/osu/master/LICENCE
+
+using osu.Framework.Allocation;
+using osu.Framework.Graphics;
+using osu.Game.Input.Bindings;
+using osu.Game.Overlays.Settings;
+using osu.Game.Rulesets;
+
+namespace osu.Game.Overlays.KeyConfiguration
+{
+    public class KeyConfiguration : SettingsOverlay
+    {
+        protected override Drawable CreateHeader() => new SettingsHeader("key configuration", "Customise your keys!");
+
+        [BackgroundDependencyLoader(permitNulls: true)]
+        private void load(RulesetStore rulesets, GlobalBindingInputManager global)
+        {
+            AddSection(new GlobalBindingsSection(global, "Global"));
+
+            foreach (var ruleset in rulesets.Query<RulesetInfo>())
+                AddSection(new RulesetBindingsSection(ruleset));
+        }
+
+        public KeyConfiguration()
+            : base(false)
+        {
+        }
+    }
+}
